+++ rust/net/src/chat.rs
@@ -23,6 +23,7 @@ use libsignal_net_infra::{
     AsHttpHeader, AsStaticHttpHeader, Connection, IpType, RECOMMENDED_WS_CONFIG, TransportInfo,
 };
 use tokio_tungstenite::WebSocketStream;
+use tungstenite::protocol::WebSocketConfig;
 
 use crate::auth::Auth;
 use crate::connect_state::{
@@ -222,6 +223,11 @@ impl ChatHeaders {
     }
 }
 
+pub enum EnablePermessageDeflate {
+    No,
+    Yes,
+}
+
 pub type ChatServiceRoute = UnresolvedWebsocketServiceRoute;
 
 impl ChatConnection {
@@ -230,6 +236,7 @@ impl ChatConnection {
         http_route_provider: impl RouteProvider<Route = UnresolvedHttpsServiceRoute>,
         user_agent: &UserAgent,
         ws_config: self::ws::Config,
+        enable_permessage_deflate: EnablePermessageDeflate,
         headers: Option<ChatHeaders>,
         log_tag: &str,
     ) -> Result<PendingChatConnection, ConnectError>
@@ -244,6 +251,7 @@ impl ChatConnection {
             http_route_provider,
             user_agent,
             ws_config,
+            enable_permessage_deflate,
             headers,
             log_tag,
         )
@@ -256,6 +264,7 @@ impl ChatConnection {
         http_route_provider: impl RouteProvider<Route = UnresolvedHttpsServiceRoute>,
         user_agent: &UserAgent,
         ws_config: self::ws::Config,
+        enable_permessage_deflate: EnablePermessageDeflate,
         headers: Option<ChatHeaders>,
         log_tag: &str,
     ) -> Result<PendingChatConnection<TC::Connection>, ConnectError>
@@ -269,8 +278,13 @@ impl ChatConnection {
             .into_iter()
             .flat_map(ChatHeaders::iter_headers)
             .chain([user_agent.as_header()]);
+        let mut ws_stream_config = WebSocketConfig::default();
+        ws_stream_config.extensions.permessage_deflate = match enable_permessage_deflate {
+            EnablePermessageDeflate::Yes => Some(Default::default()),
+            EnablePermessageDeflate::No => None,
+        };
         let ws_fragment = WebSocketRouteFragment {
-            ws_config: Default::default(),
+            ws_config: ws_stream_config,
             endpoint: PathAndQuery::from_static(crate::env::constants::WEB_SOCKET_PATH),
             headers: HeaderMap::from_iter(headers),
         };
@@ -307,6 +321,10 @@ impl ChatConnection {
             stream,
             response_headers,
         } = connection.into_inner();
+
+        if stream.get_config().extensions.permessage_deflate.is_some() {
+            log::info!("[{log_tag}] established connection is using permessage-deflate")
+        }
 
         Ok(PendingChatConnection {
             connection: stream,
@@ -469,6 +487,7 @@ pub mod test_support {
             route_provider,
             &user_agent,
             ws_config,
+            EnablePermessageDeflate::No,
             None,
             "test",
         )
@@ -739,6 +758,7 @@ pub(crate) mod test {
                 remote_idle_timeout: Duration::ZERO,
                 initial_request_id: 0,
             },
+            EnablePermessageDeflate::No,
             None,
             "fake chat",
         )
@@ -838,6 +858,7 @@ pub(crate) mod test {
                 remote_idle_timeout: Duration::ZERO,
                 initial_request_id: 0,
             },
+            EnablePermessageDeflate::No,
             Some(auth_headers.clone().into()),
             "fake chat",
         )
@@ -859,6 +880,7 @@ pub(crate) mod test {
                 remote_idle_timeout: Duration::ZERO,
                 initial_request_id: 0,
             },
+            EnablePermessageDeflate::No,
             Some(auth_headers.into()),
             "fake chat",
         )