+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.81.1
 
+- Enable negotiating permessage-deflate support for chat websocket connections, if configured.